pyproject.toml
[tool.poetry]
name = "dylan-community-be"
version = "0.1.0"
description = ""
authors = ["Hongmin Lee <107842030+webplusangels@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "~3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.35"
pymysql = "^1.1.1"
pydantic = {extras = ["email"], version = "^2.11.5"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
python-multipart = "^0.0.12"
pydantic-settings = "^2.9.1"
aiosqlite = "^0.21.0"
passlib = {version = ">=1.7.4", extras = ["bcrypt"]}
bcrypt = "4.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pre-commit = "^4.2.0"
ruff = "^0.11.12"
httpx = "^0.28.1"
pyrefly = "^0.19.0"
pytest-mock = "^3.14.1"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py313"
line-length = 88
lint.select = [
    "E",
    "W",
    "F",
    "I",
    "C",
    "B",
]
lint.ignore = [
    "E501",  # line too long (handled by formatter)
    "E402",  # module level import not at top of file (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["src"]
